requirements.txt
aiohttp==3.7.4.post0
async-timeout
attrs
chardet
discord.py
idna
multidict
Pillow
PyYAML
typing-extensions
yarl
python-dotenv